feat(eve-client): add API URL configuration and implement fetching of default character and list of characters from the API
This commit introduces a new environment variable `VITE_API_URL` to configure the URL for the API. It also adds functions to fetch the default character and a list of characters from the API using the new environment variable. The old hardcoded API URL has been removed from the client-side code. This change allows the application to be easily configured to use different APIs by simply setting the `VITE_API_URL` environment variable.